pyproject.toml
[tool.poetry]
name = "applied-data-science-for-business"
version = "0.1.0"
description = "Applied Data Science for Business - A comprehensive data science project"
authors = ["Maria Lucia Castro Jorge <mluciacjds@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "*"
pandas = "^2.0.0"
s3fs = "^2024.0.0"
pyarrow = "^15.0.0"
fastapi = "*"
uvicorn = { extras = ["standard"], version = "*" }
pydantic = "*"
pydantic-settings = "*"
scipy = "^1.10.0"
statsmodels = "^0.14.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
flake8 = "^5.0.0"
ipykernel = { version = "^6.29.0", python = ">=3.9" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
